Unify mempool configs (#3983)
# Description
Especially on gnosis chain the reference driver struggled to bring
solutions on chain. The issue seems to have been poor transaction
gossiping of the 1 configured RPC. The fix was to configure multiple
mempools at once. This ended up looking pretty weird with the current
code because the only way how you can configure an RPC URL explicitly is
by using the mevblocker variant of the mempool enum. This doesn't make
sense since a) there is no mevblocker on gnosis chain and b) even on
mainnet there is only 1 mevblocker.

# Changes
This PR removes the `mempool` enum and replaces it with a struct that is
flexible enough to configure an actual mevblocker endpoint but also
multiple regular RPC nodes.

Since this is changing the driver config file in a backwards
incompatible way this needs to be merged together with the corresponding
infra PR updating the config format.

## How to test
updated e2e and driver tests should still pass