Implement sell=buy support for Sell orders (#3894)
# Sell=buy token initiative
The ability to swap the same token pairs unlocks the potential for CoW
to generally become a transaction relay service. This feature will allow
users to execute pre- and post- interaction hooks as gasless
transactions. The overall implementation has certain caveats:
- Each settlement has to go through the settlement contract anyway, so
the sell=buy token transactions would require users to send X amount of
sell token and receive X - cost of the token back. It is not realistic,
with the current state of GPv2 contract for the user to just send the
cost amount and save on gas this way.
- The solvers can really compete only by batching to save on gas costs
- There is little incentive for this feature to be supported by other
solvers, but having this support in baseline solver could generally be
enough.

# Lift restrictions on sell=buy token
> Based the changes on
https://github.com/cowprotocol/services/pull/2513/files, adapted to
current state of the codebase.

- Adds a command-line argument to the orderbook API to allow for same
sell and buy token swaps. The SameBuyAndSell token error is not raised
in that case
- Adds e2e test case that checks if sell=buy quoting and orders are
possible
- Liquidity fetching happens only if the sell and buy tokens are
different
- Relaxes the invariant on Tokens struct to allow for same token pairs
- Minor tweak to Justfile to allow for custom filtering and extra args
to nextest

Based on input from @fhenneke:
- Adds a short-circuit behavior for solver to return empty route in case
sell=buy
- Returns sell and buy assets directly when solving for empty route

---------

Co-authored-by: ilya <ilya@cow.fi>
Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>
Co-authored-by: José Duarte <duarte.gmj@gmail.com>