feat(ci): add check to ensure artifacts are in sync with contract sources (#3991)
# Description
<!--- Describe your changes to provide context for reviewers, including
why it is needed -->

On my other PR https://github.com/cowprotocol/services/pull/3937 , I was
having difficulty generating the solidity JSONs, and when I did generate
them, the format was not correct. I realized that part of the problem is
that the CI should be natively checking for this, but it currently does
not, so I added a job.

Additionally, my computer is ARM64, but the ethereum solidity image is
only compatible with `x86_64`, so I had to add a flag to fix this.

Finally, the format followed by the other files in the artifacts JSON
directory uses a pretty-printed JSON format, but the existing Makefile
uses compact format with `jq -c`. So I removed the `-c` flag to ensure
this consistent format.

It turns out taht, likely due to a newer solidity version, many of the
artifacts have changed slightly since they were originally built. So
that is included as part of this PR.

One other issue that can occur is if the artifact file already exists on
your computer, it may fail to write the new version since `make` only
builds files that don't exist (or something like that). To resolve this,
the `-B` option is used to force rebuilding of all targets.

# Changes
<!-- List of detailed changes (how the change is accomplished) -->

- [x] modify the Makefile as described above
- [x] add CI job
- [x] Ensure the CI job can run properly on ARM64
- [x] run `cd crates/contracts/solidity && make -B artifacts`

## How to test

Check the new CI job `solidity-artifacts`

Run `make -B artifacts` from `crates/contracts/solidity` to see how it
behaves on your own machine.

<!--
## Related Issues

Fixes #
-->