feat(balance_overrides): use debug_traceCall to find slots (#3937)
# Description

Implement a more reliable, and likely (usually) faster approach for
detecting solidity balance override storage addresses.

Rather than bulk storage override/scanning many storage slots for a
match, a single `debug_traceCall` is made on the ERC20 `balanceOf`
function, and reading the resolved `SLOAD` slots. If only one `SLOAD` is
detected, then we are done. If more than one slot is found, each slot
can be tested through `eth_call` with a single slot override. We test
from last accessed to first accessed, as the last accessed storage slot
is theoretically the most likely to be the balance slot (as that would
be the return value). Since storage slots could be ccessed from anywhere
(ex. not just in a solidity mapping), a new `DirectSlot` strategy has
been added.

Probably the only major drawback of the `DirectSlot` strategy is that it
can only be used for a single account. If a separate account balance
from the check needs to be overridden, either the slot has to be
recalculated, or the balance has to be sent from a separate address
after allocation (ex. Spardose). Due to this limitation, any detected
slots are additionally tested to match up with `SolidityMapping`, and
the `SolidityMapping` strategy is returned instead of `DirectSlot` if
able.

This approach is very similar to [one followed by Foundry for their
`deal`
cheatcode](https://github.com/foundry-rs/foundry/blob/9b13b811849e73654fae046986b8730df8a0d64d/crates/anvil/src/eth/api.rs#L2259).
This is replacing the existing pattern of overriding 50 storage slots
and seeing which test balance gets applied.

This method has advantages:
* Assuming that there is only one `SLOAD` in a `balanceOf` call (most
cases), only a single `debug_traceCall` is required for a single
token/address pair. In the case of more than one SLOAD, an additional
call is required for each slot to validate the correct storage, which is
still inexpensive.
* Works even if the token calls a separate contract to get the balance
(rather than reading from its internal storage)
* Basically any token that stores a user's balance in a single `uint256`
will be supported by this method.
* Since this strategy can be reliably used anywhere, we can potentially.
Performance allowing.

Potential drawbacks:
* `debug_traceCall` isnt supported everywhere. This could be overcome by
using `eth_createAccessList` as a fallback, but the createAccessList API
does not return accessed storage slots in order, so the check could be
considerably more expensive if the balance function accesses many slots.
* The `DirectSlot` strategy can only update the balance of a single
token holder whereas `SolidityMapping` can update any holder.


Other considerations:
* Since we now support just about every reasonable token out there with
this system, we should consider automatically banning any tokens for
which we cannot override the balance. this will make simulations much
more simple and reliable.

Since slot detection using this method is only able to find the storage
slot for a specific account in question, the interface needed to be
updated in a couple places to reflect this. This also means that there
is a potential performance disadvantage with this method since balance
overrides for different addresses cannot be detected due to not
computing via Solidity mapping. This could be mitigated by only
supporting balance overrides on the spardose contract (followed by a
`transfer` call) to the needed account as required.

This method requires a node that supports the `debug_traceCall` API. I
believe this is the case for our infra, but please double check :pray:.

When possible this storage detection method will use `SolidityMapping`
to assist in finding cases where it is possible to use. So there should
be no need to do heuristic scanning anymore, so I have removed this code
segment (and any supporting code).

I added a E2E test and contract to validate that the storage slot
detection is working as expected.

# Changes


- [x] add `DirectSlot` detector
- [x] update caching interface to cache by the pair (token address,
overridden balance address)
- [x] add E2E test to verify the detector works in practice

## How to test

Run `just test-e2e local_node_trace_based_balance_detection`

---------

Co-authored-by: Jan [Yann] <4518474+fafk@users.noreply.github.com>
Co-authored-by: ilya <ilya@cow.fi>
Co-authored-by: José Duarte <duarte.gmj@gmail.com>