Migrate RPC calls into alloy (#3982)
# Description
Migrates almost all RPC calls to the alloy provider. As far as I can
tell, the only remaining ones are related to the access lists which
require more attention.

There are others missing because they use the `transport` variable and I
was only searching for `web3`; I'll leave it for another PR

# Changes

- [ ] Migrate existing RPC calls into alloy
- [ ] Refactor where applicable


## How to test
Existing tests + staging