Simulate solution on pending block (#3992)
# Description
When the driver receives a solution from the solver it re-simulates it
to make sure it doesn't actually revert.
Additionally the driver re-simulates a solution right before submitting
it to the chain (in case it turned sour while the driver waited to be
crowned the winner of the auction).
Both of these checks currently happen on the `latest` block but that is
actually more risky than necessary. Instead we can also simulate on the
`pending` block. This effectively tells the node to simulate the
transaction taking any higher priced transactions into account. If there
is already a transaction in the mempool that tries to take advantage of
the same arbitrage opportunity as our solution but is higher priced than
our tx the solution would revert on the `pending` block but not on
`latest`.
Since we error out as early as possible to prevent the case of winning
an auction with a solution that is doomed to fail it makes more sense to
simulate on the `pending` block.

# Changes
Changed driver code to do simulations on the `pending` instead of
`latest` block

## How to test
Since this is treating a bit of an edge case and relies on actual prod
traffic it would likely take quite a while running this in prod to
clearly see the effects. I suggest to merge as is and re-evaluate after
having this on all networks for a week.