Remove soft cancellation (#3993)
# Description
Removes the `use-soft-cancellation` flag from the mempool configs. This
flag is currently not used to make any decisions. And more importantly
it will actually never make sense to execute different logic based on
that.
The reason is that we actually don't care about cancellations getting
mined. If we submit a cancellation we are really only interested in
preventing the previous settlement tx to no longer be in the mempool.
If the same tx got replaced many times over it's technically possible
that it's more cost effective to have an already expensive cancellation
tx get mined than replacing it with another settlement tx (making it
even more expensive in the process) but given the bad state of our
current submission logic this is nothing we should optimize for at the
moment.

# Changes
Deleted unused `use_soft_cancellations` flag