Simplify tx gas price logic (#3994)
# Description
Applies simplifications and savings to the tx gas price logic.

# Changes
Simplifications:
* don't take a `deadline` argument in the `gas_price()` function anymore
since it was always `None`
* stop fetching the block in the alloy gas price estimator since it's
only used for the `base_fee` which doesn't make sense in the context we
need gas price estimation anyway
* use `GasPriceEstimating::estimate()` instead of
`GasPriceEstimating::estimate_with_limits()` because we always used the
defaults anyway

Savings:
Rely on gas price estimator to provide a reasonable gas price instead of
bumping the gas price once per block until the submissions deadline. The
alloy gas price estimator that we use actually already returns a
`max_fee_per_gas` that is twice as high as the current base fee (see
[here](https://github.com/alloy-rs/alloy/blob/main/crates/provider/src/utils.rs#L116))
which already be sufficient for the tight deadlines we now have.
Additionally the logic that increased the gas price for every block
until the deadline (`price * PRICE_BUMP.powi(blocks_until_deadline)`)
was only ever intended to make sure the transaction has a high enough
`max_fee_per_gas` to be executable in n blocks time. However the
multiplication also multiplied the `max_priority_fee_per_gas` which
caused us to pay significantly more even if the tx gets mined in the
very next block. (Note that multiplying both values is correct for the
case where we have to submit a cancellation because the node requires
`max_fee_per_gas` and `max_priority_fee_per_gas` to have increased to
accept the replacement).


## How to test
Briefly tested in on mainnet (most of our volume) and bnb (very costly
l2). On mainnet the tx fee didn't change much which was to be expected
because the repeated multiplying the tx gas price doesn't matter too
much there because the tx submission deadline is so short => only a few
gas price bumps.
On BNB it reduced the gas cost per tx from 1.9$ to 3ct. 🙈 
There was no increased submission delay for those very few transactions
I did.