Migrate evm_ (impersonate, automine, etc) APIs to alloy  (#4004)
# Description
Discussed in this thread:
https://github.com/cowprotocol/services/pull/4001#discussion_r2638451923

This PR moves the impersonation mechanisms to alloy, the migrated ones
are only impersonated in the specific transaction, no changes to the
actual tests were needed.

Exposed the previous string oriented API under a trait to as discussed
in a comment below.
Anvil provides an anvil specific API that we're not using unless
necessary as these ones are more general

# Changes

- [ ] Migrate the last `mine_pending_block`
- [ ] Migrate the other evm calls to the new trait
- [ ] Migrate the last impersonations
- [ ] Remove the forked_node and local_node modules

## How to test
Existing tests